refactor: extract Gist edit-session state and focus handling
